Modified UI to use GridView
Instead of implementing what boils down to be a custom adapter for
TableLayout the code has now been restructured to use the GridView
widget. A custom adapter renders the grid view from the underlying grid.
All state info is stored in the underlying grid.

The GuiGrid class has now been simplified to be a GridController class
that controlls the execution of the game (as far as managing the
grid/clues are concerned).