Added Game Chooser
Added a game chooser that allows the user to choose the game to be
played. The chooser first scans the game directory for game files and
parses only the headers of the files for game info. It then shows the
game info to the user in a list. When the user selects a game it fires
up the grid activity to let the user play the game.

With this the basic game play is complete - now the UI can be polished.